Renames
renamed executable to car-datadisplay and Project.cpp/h to main.cpp/h
